Rename ReactServerFormatConfig to ReactFizzConfig
