+++ packages/react-server/src/forks/ReactFizzConfig.dom-browser.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from 'react-dom-bindings/src/server/ReactDOMServerFormatConfig';
+export * from 'react-dom-bindings/src/server/ReactFizzConfigDOM';
